ticket-agency-app/requirements.txt
streamlit
gspread
oauth2client
pandas
numpy
python-dotenv
requests
google-api-python-client
google-auth-oauthlib
google-auth